Packageify octohat and use octohub logic inline
This is probably not the best solution, but it appears to work in testpypi (0.1.31)

Moving the octohub logic to the same level as the octohat logic, and changing the import names appears to work.
