+++ octohat/__init__.py
@@ -2,8 +2,7 @@
 
 import argparse
 import sys
-
-from utils import * 
+from helpers import * 
 
 def main():
