+++ octohat/exceptions.py
