--- octohub/octohub/__init__.py
@@ -1,12 +1,0 @@
-# Copyright (c) 2013 Alon Swartz <alon@turnkeylinux.org>
-#
-# This file is part of OctoHub.
-#
-# OctoHub is free software; you can redistribute it and/or modify it under the
-# terms of the GNU General Public License as published by the Free Software
-# Foundation; either version 3 of the License, or (at your option) any later
-# version.
-
-__version__ = '0.1'
-__useragent__ = 'octohub/%s' % __version__
-
